[FIX] Animation shared state for resources
