+++ Gameplay/entities/crystals/CrystalExplosion.cpp
@@ -8,7 +8,6 @@
 
 // TODO: These two includes must go:
 #include <modules/ModuleSceneManager.h>
-#include <resources/ResourceAnimation.h>
 
 
 Hachiko::Scripting::CrystalExplosion::CrystalExplosion(GameObject* game_object)
@@ -55,11 +54,11 @@ void Hachiko::Scripting::CrystalExplosion::OnStart()
 
 void Hachiko::Scripting::CrystalExplosion::OnUpdate()
 {
-	ResourceAnimation* resource_animation = _explosion_crystal->GetComponent<ComponentAnimation>()->GetCurrentAnimation();
+	ComponentAnimation* component_anim = _explosion_crystal->GetComponent<ComponentAnimation>();
 
-	if (!_stats->IsAlive() && resource_animation)
+	if (!_stats->IsAlive() && component_anim)
 	{
-		if (resource_animation->GetCurrentState() == ResourceAnimation::State::STOPPED) 
+		if (component_anim->AnimationIsStopped())
 		{
 			SceneManagement::Destroy(game_object);
 			return;
