+++ Gameplay/entities/enemies/EnemyController.cpp
@@ -102,8 +102,7 @@ void Hachiko::Scripting::EnemyController::OnUpdate()
 		}
 		else
 		{
-
-			if (animation->GetCurrentAnimation() != nullptr && animation->GetCurrentAnimation()->GetCurrentState() == ResourceAnimation::State::STOPPED)
+			if (animation != nullptr && animation->AnimationIsStopped())
 			{
 				DropParasite();
 			}
