+++ Source/src/components/ComponentAnimation.cpp
@@ -98,7 +98,7 @@ void Hachiko::ComponentAnimation::UpdatedGameObject(GameObject* go)
     float3 position;
     Quat rotation;
 
-    if (controller->GetTransform(controller->current, go->name.c_str(), position, rotation))
+    if (controller->GetTransform(controller->GetCurrentInstance(), go->name.c_str(), position, rotation))
     {
         go->GetTransform()->SetLocalPosition(position);
         go->GetTransform()->SetLocalRotation(rotation);
@@ -129,7 +129,7 @@ void Hachiko::ComponentAnimation::PlayNode(unsigned int node_idx, unsigned int b
 
         if (clip_idx < state_machine->GetNumClips())
         {
-            current_animation = state_machine->GetClipRes(clip_idx);
+            ResourceAnimation* current_animation = state_machine->GetClipRes(clip_idx);
 
             if (current_animation != 0)
             {
