+++ Source/src/components/ComponentAnimation.h
@@ -2,10 +2,11 @@
 
 #include "Component.h"
 
+#include "animation/AnimationController.h"
+
 namespace Hachiko
 {
     class GameObject;
-    class AnimationController;
     class ResourceAnimation;
     class ResourceStateMachine;
 
@@ -29,9 +30,9 @@ namespace Hachiko
         void Save(YAML::Node& node) const override;
         void Load(const YAML::Node& node) override;
 
-        ResourceAnimation* GetCurrentAnimation()
+        bool AnimationIsStopped()
         {
-            return current_animation;
+            return controller->GetCurrentState() == AnimationController::State::STOPPED;
         }
 
     private:
@@ -46,7 +47,6 @@ namespace Hachiko
     private:
         AnimationController* controller = nullptr;
         unsigned int active_node = 0;
-        ResourceAnimation* current_animation = nullptr;
 
         // SM CONTROL
         WindowStateMachine* windowStateMachine = nullptr;
