+++ Source/src/components/animation/AnimationController.h
@@ -10,6 +10,16 @@ namespace Hachiko
 {
     class AnimationController
     {
+    public:
+
+        enum class State
+        {
+            PLAYING, // While the animation is playing
+            PAUSED, // If the animation is paused
+            STOPPED, // When the animation is finished
+            UNSET
+        };
+
         struct Instance
         {
             ResourceAnimation* current_animation = nullptr;
@@ -33,9 +43,24 @@ namespace Hachiko
 
         bool GetTransform(Instance* instance, const std::string& channel_name, math::float3& position, Quat& rotation) const;
 
-        Instance* current = nullptr; // move to private
+        Instance* GetCurrentInstance() 
+        {
+            return current;
+        }
+
+        State GetCurrentState() const
+        {
+            return current_state;
+        }
+
+        void SetCurrentState(State new_current_state)
+        {
+            current_state = new_current_state;
+        }
 
     private:
+        Instance* current = nullptr; // move to private
+
         // INTERPOLATION
         float Interpolate(float first, float second, float lambda);
         float3 Interpolate(float3& first, float3& second, float lambda) const; 
@@ -44,5 +69,7 @@ namespace Hachiko
         // MANAGE INSTANCES
         void UpdateInstance(Instance* instance, unsigned elapsed);
         void ReleaseInstance(Instance* instance);
+
+        State current_state = State::UNSET;
     };
 } // namespace Hachiko
